+++ idea/testData/quickfix/autoImports/noImportInQualifiedUserTypeNotFirst.before.Main.kt
@@ -1,5 +1,6 @@
 // "class org.jetbrains.kotlin.idea.quickfix.AutoImportFix" "false"
 // ACTION: Create interface 'SomeTest'
+// ACTION: Create type alias 'SomeTest'
 // ERROR: Unresolved reference: SomeTest
 
 package testing