+++ idea/testData/quickfix/createFromUsage/createClass/typeReference/enumEntryJavaEnumReceiver.before.Main.kt
@@ -5,5 +5,6 @@
 // ACTION: Create class 'A'
 // ACTION: Create enum 'A'
 // ACTION: Create interface 'A'
+// ACTION: Create type alias 'A'
 // ERROR: Unresolved reference: A
 internal fun foo(): J.<caret>A = throw Throwable("")