+++ idea/testData/quickfix/createFromUsage/createTypeAlias/typeReference/currentPackageReceiver.kt
@@ -1,0 +1,5 @@
+// "Create type alias 'A'" "true"
+// ERROR: Unresolved reference: Dummy
+package p
+
+fun foo(): p.<caret>A = throw Throwable("")