+++ idea/testData/quickfix/createFromUsage/createTypeAlias/typeReference/nestedClassByTypeArgumentRefNoBody.kt
@@ -1,0 +1,4 @@
+// "Create type alias 'X'" "true"
+open class A<T>
+
+class B : A<B.<caret>X>()