+++ idea/testData/quickfix/createFromUsage/createTypeAlias/typeReference/nestedClassByTypeArgumentRefWithBody.kt
@@ -1,0 +1,6 @@
+// "Create type alias 'X'" "true"
+open class A<T>
+
+class B : A<B.<caret>X>() {
+
+}