+++ idea/testData/quickfix/createFromUsage/createTypeAlias/typeReference/userTypeReceiver.kt
@@ -1,0 +1,9 @@
+// "Create type alias 'A'" "true"
+// ERROR: Unresolved reference: Dummy
+package p
+
+class T {
+
+}
+
+fun foo(): T.<caret>A = throw Throwable("")